REFACTOR(ViewModel): Replace RemoteViewModel with ImageViewModel.
- Removed `RemoteViewModel` from `ui.home`, as it was handling nurse images and API calls inefficiently.

- Created `ImageViewModel` in a new package `image` to better separate concerns.

- Improved state management using `NursesUiState` to handle loading, success, and error states.

- Ensured `fetchNurses()` initializes automatically when `ImageViewModel` is instantiated.

- Optimized `getNurseImage()` method for better error handling and result validation.
